# docs/swagger.yaml
basePath: /
definitions:
  handlers.createPersonRequest:
    properties:
      age:
        type: integer
      name:
        type: string
    type: object
  handlers.updatePersonRequest:
    properties:
      age:
        type: integer
      name:
        type: string
    type: object
  models.Person:
    properties:
      age:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: RESTful API for managing people
  title: PeopleHub API
  version: "1.0"
paths:
  /people:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Person'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List all people
      tags:
      - People
    post:
      consumes:
      - application/json
      parameters:
      - description: Person JSON
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/handlers.createPersonRequest'
      produces:
      - text/plain
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Add a new person
      tags:
      - People
  /people/{id}:
    delete:
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete person by ID
      tags:
      - People
    get:
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Person'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get person by ID
      tags:
      - People
    patch:
      consumes:
      - application/json
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/handlers.updatePersonRequest'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update person fields
      tags:
      - People
  /people/age:
    get:
      parameters:
      - description: Minimum age
        in: query
        name: min
        required: true
        type: integer
      - description: Maximum age
        in: query
        name: max
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Person'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get people by age range
      tags:
      - People
  /people/recent:
    get:
      parameters:
      - description: Limit number of people
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Person'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get recent people
      tags:
      - People
  /people/search:
    get:
      parameters:
      - description: Name keyword
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Person'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Search people by name
      tags:
      - People
  /people/stats:
    get:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get total number of people
      tags:
      - People
swagger: "2.0"